src/requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
asttokens==2.0.4
attrs==21.2.0
billiard==3.6.3.0
CacheControl==0.12.10
cachetools==4.2.4
celery==4.4.7
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.0rc2
Django==3.1.13
django-braces==1.14.0
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cors==0.1
django-cors-headers==3.9.0
django-debug-toolbar==3.2
django-filter==2.4.0
django-oauth-toolkit==1.4.0
django-rest-framework-social-oauth2==1.1.0
django-safedelete==0.5.7
django-shortuuidfield==0.1.3
django-singleton-admin==0.0.4
django-templated-mail==1.1.1
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
executing==0.5.4
fcm-django==1.0.8
firebase-admin==5.2.0
flower==0.9.7
google-api-core==2.0.1
google-api-python-client==2.23.0
google-auth==2.2.1
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.53.0
grpcio==1.44.0
httplib2==0.19.1
humanize==3.2.0
hurry==1.1
hurry.filesize==0.9
icecream==2.1.0
idna==2.10
inflection==0.5.1
isort==5.7.0
itypes==1.2.0
Jinja2==2.11.3
jsonschema==4.0.1
kombu==4.6.11
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.3
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==20.9
pathspec==0.8.1
Pillow==8.3.2
prometheus-client==0.8.0
prompt-toolkit==3.0.14
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.4
PyJWT==2.0.1
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.16.12
shortuuid==1.0.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
vine==1.3.0
wcwidth==0.2.5
wrapt==1.12.1
